FixedSomething

+++ src/Services/Job/JobCompany/JobCompany.Business/Dtos/ExamDtos/CreateExamDto.cs
@@ -8,7 +8,7 @@ namespace JobCompany.Business.Dtos.ExamDtos
         public string? Title { get; set; }
         public string? IntroDescription { get; set; }
         public string? LastDescription { get; set; }
-        public decimal LimitRate { get; set; }
+        public float LimitRate { get; set; }
         public ICollection<QuestionCreateDto>? Questions { get; set; }
         public bool IsTemplate { get; set; }
         public short? Duration { get; set; }
